+++ test/unit/strats.unit.test.ts
@@ -2,12 +2,39 @@ import assert from "assert";
 import { describe, it } from "mocha";
 
 import Mangrove_v2_0_1 from "../../src/assets/core/v2.0.1/Mangrove.json";
+import BlastMangrove_v2_1_0_0 from "../../src/assets/core/v2.1.0-0/BlastMangrove.json";
+// v1.0.0
 import AaveKandelSeeder_v1_0_0 from "../../src/assets/strats/v1.0.0/AaveKandelSeeder.json";
 import AavePooledRouter_v1_0_0 from "../../src/assets/strats/v1.0.0/AavePooledRouter.json";
 import KandelLib_v1_0_0 from "../../src/assets/strats/v1.0.0/KandelLib.json";
 import KandelSeeder_v1_0_0 from "../../src/assets/strats/v1.0.0/KandelSeeder.json";
 import MangroveOrderRouter_v1_0_0 from "../../src/assets/strats/v1.0.0/MangroveOrder-Router.json";
 import MangroveOrder_v1_0_0 from "../../src/assets/strats/v1.0.0/MangroveOrder.json";
+// v2.0.0-b1.0
+import AaveKandelSeeder_v2_0_0_b1_0 from "../../src/assets/strats/v2.0.0-b1.0/AaveKandelSeeder.json";
+import AavePooledRouter_v2_0_0_b1_0 from "../../src/assets/strats/v2.0.0-b1.0/AavePooledRouter.json";
+import KandelLib_v2_0_0_b1_0 from "../../src/assets/strats/v2.0.0-b1.0/KandelLib.json";
+import KandelSeeder_v2_0_0_b1_0 from "../../src/assets/strats/v2.0.0-b1.0/KandelSeeder.json";
+import MangroveOrderRouter_v2_0_0_b1_0 from "../../src/assets/strats/v2.0.0-b1.0/MangroveOrder-Router.json";
+import MangroveOrder_v2_0_0_b1_0 from "../../src/assets/strats/v2.0.0-b1.0/MangroveOrder.json";
+import RouterProxyFactory_v2_0_0_b1_0 from "../../src/assets/strats/v2.0.0-b1.0/RouterProxyFactory.json";
+import SimpleAaveLogic_v2_0_0_b1_0 from "../../src/assets/strats/v2.0.0-b1.0/SimpleAaveLogic.json";
+import MangroveAmplifier_v2_0_0_b1_0 from "../../src/assets/strats/v2.0.0-b1.0/MangroveAmplifier.json";
+// v2.0.1-0
+import AaveKandelSeeder_v2_0_1_0 from "../../src/assets/strats/v2.0.1-0/AaveKandelSeeder.json";
+import AavePooledRouter_v2_0_1_0 from "../../src/assets/strats/v2.0.1-0/AavePooledRouter.json";
+import KandelLib_v2_0_1_0 from "../../src/assets/strats/v2.0.1-0/KandelLib.json";
+import KandelSeeder_v2_0_1_0 from "../../src/assets/strats/v2.0.1-0/KandelSeeder.json";
+import MangroveAmplifier_v2_0_1_0 from "../../src/assets/strats/v2.0.1-0/MangroveAmplifier.json";
+import MangroveOrderRouter_v2_0_1_0 from "../../src/assets/strats/v2.0.1-0/MangroveOrder-Router.json";
+import MangroveOrder_v2_0_1_0 from "../../src/assets/strats/v2.0.1-0/MangroveOrder.json";
+import RouterProxyFactory_v2_0_1_0 from "../../src/assets/strats/v2.0.1-0/RouterProxyFactory.json";
+import SimpleAaveLogic_v2_0_1_0 from "../../src/assets/strats/v2.0.1-0/SimpleAaveLogic.json";
+// v2.1.0-0
+import BlastMangroveAmplifier_v2_1_0_0 from "../../src/assets/strats/v2.1.0-0/BlastMangroveAmplifier.json";
+import BlastMangroveOrderRouter_v2_1_0_0 from "../../src/assets/strats/v2.1.0-0/BlastMangroveOrder-Router.json";
+import BlastMangroveOrder_v2_1_0_0 from "../../src/assets/strats/v2.1.0-0/BlastMangroveOrder.json";
+import BlastRouterProxyFactory_v2_1_0_0 from "../../src/assets/strats/v2.1.0-0/BlastRouterProxyFactory.json";
 
 import {
   getAaveKandelSeederVersionDeployments,
@@ -18,6 +45,9 @@ import {
   getAllKandelSeederVersionDeploymentsPerNetwork,
   getAllMangroveOrderRouterVersionDeploymentsPerNetwork,
   getAllMangroveOrderVersionDeploymentsPerNetwork,
+  getAllRouterProxyFactoryVersionDeploymentsPerNetwork,
+  getAllSimpleAaveLogicVersionDeploymentsPerNetwork,
+  getAllMangroveAmplifierVersionDeploymentsPerNetwork,
   getKandelLibVersionDeployments,
   getKandelSeederVersionDeployments,
   getLatestAaveKandelSeederPerNetwork,
@@ -26,9 +56,15 @@ import {
   getLatestKandelSeederPerNetwork,
   getLatestMangroveOrderPerNetwork,
   getLatestMangroveOrderRouterPerNetwork,
+  getLatestRouterProxyFactoryPerNetwork,
+  getLatestSimpleAaveLogicPerNetwork,
   getLatestStratContractsPerNetwork,
+  getLatestMangroveAmplifierPerNetwork,
   getMangroveOrderRouterVersionDeployments,
   getMangroveOrderVersionDeployments,
+  getRouterProxyFactoryVersionDeployments,
+  getSimpleAaveLogicVersionDeployments,
+  getMangroveAmplifierVersionDeployments,
 } from "../../src/strats";
 import { expect } from "chai";
 import { firstVersionDeploymentsToVersionNetworkDeployment } from "./unitTestUtil";
@@ -40,11 +76,13 @@ describe("strats.ts", () => {
         const result = getAaveKandelSeederVersionDeployments({
           released: undefined,
         });
-        assert.equal(result, AaveKandelSeeder_v1_0_0);
+        assert.equal(result, AaveKandelSeeder_v2_0_1_0);
         // NB: Add older old versions here
-        [].forEach((version) => {
-          assert.notEqual(result, version);
-        });
+        [AaveKandelSeeder_v2_0_0_b1_0, AaveKandelSeeder_v1_0_0].forEach(
+          (version) => {
+            assert.notEqual(result, version);
+          },
+        );
       });
     });
 
@@ -55,7 +93,11 @@ describe("strats.ts", () => {
             released: undefined,
           }),
         ).to.deep.equal({
-          "80001": [AaveKandelSeeder_v1_0_0],
+          "80001": [
+            AaveKandelSeeder_v2_0_1_0,
+            AaveKandelSeeder_v2_0_0_b1_0,
+            AaveKandelSeeder_v1_0_0,
+          ],
         });
       });
     });
@@ -66,7 +108,7 @@ describe("strats.ts", () => {
           getLatestAaveKandelSeederPerNetwork({ released: undefined }),
         ).to.deep.equal({
           "80001": firstVersionDeploymentsToVersionNetworkDeployment(
-            AaveKandelSeeder_v1_0_0,
+            AaveKandelSeeder_v2_0_1_0,
             "80001",
           ),
         });
@@ -80,11 +122,13 @@ describe("strats.ts", () => {
         const result = getAavePooledRouterVersionDeployments({
           released: undefined,
         });
-        assert.equal(result, AavePooledRouter_v1_0_0);
+        assert.equal(result, AavePooledRouter_v2_0_1_0);
         // NB: Add older old versions here
-        [].forEach((version) => {
-          assert.notEqual(result, version);
-        });
+        [AavePooledRouter_v2_0_0_b1_0, AavePooledRouter_v1_0_0].forEach(
+          (version) => {
+            assert.notEqual(result, version);
+          },
+        );
       });
     });
 
@@ -95,7 +139,11 @@ describe("strats.ts", () => {
             released: undefined,
           }),
         ).to.deep.equal({
-          "80001": [AavePooledRouter_v1_0_0],
+          "80001": [
+            AavePooledRouter_v2_0_1_0,
+            AavePooledRouter_v2_0_0_b1_0,
+            AavePooledRouter_v1_0_0,
+          ],
         });
       });
     });
@@ -106,7 +154,7 @@ describe("strats.ts", () => {
           getLatestAavePooledRouterPerNetwork({ released: undefined }),
         ).to.deep.equal({
           "80001": firstVersionDeploymentsToVersionNetworkDeployment(
-            AavePooledRouter_v1_0_0,
+            AavePooledRouter_v2_0_1_0,
             "80001",
           ),
         });
@@ -118,9 +166,9 @@ describe("strats.ts", () => {
     describe("getKandelLibVersionDeployments", () => {
       it("should find the latest deployment first", () => {
         const result = getKandelLibVersionDeployments({ released: undefined });
-        assert.equal(result, KandelLib_v1_0_0);
+        assert.equal(result, KandelLib_v2_0_1_0);
         // NB: Add older old versions here
-        [].forEach((version) => {
+        [KandelLib_v2_0_0_b1_0, KandelLib_v1_0_0].forEach((version) => {
           assert.notEqual(result, version);
         });
       });
@@ -131,7 +179,12 @@ describe("strats.ts", () => {
         expect(
           getAllKandelLibVersionDeploymentsPerNetwork({ released: undefined }),
         ).to.deep.equal({
-          "80001": [KandelLib_v1_0_0],
+          "80001": [
+            KandelLib_v2_0_1_0,
+            KandelLib_v2_0_0_b1_0,
+            KandelLib_v1_0_0,
+          ],
+          "168587773": [KandelLib_v2_0_1_0],
         });
       });
     });
@@ -142,8 +195,12 @@ describe("strats.ts", () => {
           getLatestKandelLibPerNetwork({ released: undefined }),
         ).to.deep.equal({
           "80001": firstVersionDeploymentsToVersionNetworkDeployment(
-            KandelLib_v1_0_0,
+            KandelLib_v2_0_1_0,
             "80001",
+          ),
+          "168587773": firstVersionDeploymentsToVersionNetworkDeployment(
+            KandelLib_v2_0_1_0,
+            "168587773",
           ),
         });
       });
@@ -156,9 +213,9 @@ describe("strats.ts", () => {
         const result = getKandelSeederVersionDeployments({
           released: undefined,
         });
-        assert.equal(result, KandelSeeder_v1_0_0);
+        assert.equal(result, KandelSeeder_v2_0_1_0);
         // NB: Add older old versions here
-        [].forEach((version) => {
+        [KandelSeeder_v2_0_0_b1_0, KandelSeeder_v1_0_0].forEach((version) => {
           assert.notEqual(result, version);
         });
       });
@@ -171,7 +228,12 @@ describe("strats.ts", () => {
             released: undefined,
           }),
         ).to.deep.equal({
-          "80001": [KandelSeeder_v1_0_0],
+          "80001": [
+            KandelSeeder_v2_0_1_0,
+            KandelSeeder_v2_0_0_b1_0,
+            KandelSeeder_v1_0_0,
+          ],
+          "168587773": [KandelSeeder_v2_0_1_0],
         });
       });
     });
@@ -182,8 +244,12 @@ describe("strats.ts", () => {
           getLatestKandelSeederPerNetwork({ released: undefined }),
         ).to.deep.equal({
           "80001": firstVersionDeploymentsToVersionNetworkDeployment(
-            KandelSeeder_v1_0_0,
+            KandelSeeder_v2_0_1_0,
             "80001",
+          ),
+          "168587773": firstVersionDeploymentsToVersionNetworkDeployment(
+            KandelSeeder_v2_0_1_0,
+            "168587773",
           ),
         });
       });
@@ -196,9 +262,13 @@ describe("strats.ts", () => {
         const result = getMangroveOrderRouterVersionDeployments({
           released: undefined,
         });
-        assert.equal(result, MangroveOrderRouter_v1_0_0);
+        assert.equal(result, BlastMangroveOrderRouter_v2_1_0_0);
         // NB: Add older old versions here
-        [].forEach((version) => {
+        [
+          MangroveOrderRouter_v2_0_1_0,
+          MangroveOrderRouter_v2_0_0_b1_0,
+          MangroveOrderRouter_v1_0_0,
+        ].forEach((version) => {
           assert.notEqual(result, version);
         });
       });
@@ -211,7 +281,12 @@ describe("strats.ts", () => {
             released: undefined,
           }),
         ).to.deep.equal({
-          "80001": [MangroveOrderRouter_v1_0_0],
+          "80001": [
+            MangroveOrderRouter_v2_0_1_0,
+            MangroveOrderRouter_v2_0_0_b1_0,
+            MangroveOrderRouter_v1_0_0,
+          ],
+          "168587773": [BlastMangroveOrderRouter_v2_1_0_0],
         });
       });
     });
@@ -222,8 +297,12 @@ describe("strats.ts", () => {
           getLatestMangroveOrderRouterPerNetwork({ released: undefined }),
         ).to.deep.equal({
           "80001": firstVersionDeploymentsToVersionNetworkDeployment(
-            MangroveOrderRouter_v1_0_0,
+            MangroveOrderRouter_v2_0_1_0,
             "80001",
+          ),
+          "168587773": firstVersionDeploymentsToVersionNetworkDeployment(
+            BlastMangroveOrderRouter_v2_1_0_0,
+            "168587773",
           ),
         });
       });
@@ -236,9 +315,13 @@ describe("strats.ts", () => {
         const result = getMangroveOrderVersionDeployments({
           released: undefined,
         });
-        assert.equal(result, MangroveOrder_v1_0_0);
+        assert.equal(result, BlastMangroveOrder_v2_1_0_0);
         // NB: Add older old versions here
-        [].forEach((version) => {
+        [
+          MangroveOrder_v2_0_1_0,
+          MangroveOrder_v2_0_0_b1_0,
+          MangroveOrder_v1_0_0,
+        ].forEach((version) => {
           assert.notEqual(result, version);
         });
       });
@@ -251,7 +334,12 @@ describe("strats.ts", () => {
             released: undefined,
           }),
         ).to.deep.equal({
-          "80001": [MangroveOrder_v1_0_0],
+          "80001": [
+            MangroveOrder_v2_0_1_0,
+            MangroveOrder_v2_0_0_b1_0,
+            MangroveOrder_v1_0_0,
+          ],
+          "168587773": [BlastMangroveOrder_v2_1_0_0],
         });
       });
     });
@@ -262,8 +350,149 @@ describe("strats.ts", () => {
           getLatestMangroveOrderPerNetwork({ released: undefined }),
         ).to.deep.equal({
           "80001": firstVersionDeploymentsToVersionNetworkDeployment(
-            MangroveOrder_v1_0_0,
+            MangroveOrder_v2_0_1_0,
             "80001",
+          ),
+          "168587773": firstVersionDeploymentsToVersionNetworkDeployment(
+            BlastMangroveOrder_v2_1_0_0,
+            "168587773",
+          ),
+        });
+      });
+    });
+  });
+
+  describe("RouterProxyFactory contract", () => {
+    describe("getRouterProxyFactoryVersionDeployments", () => {
+      it("should find the latest deployment first", () => {
+        const result = getRouterProxyFactoryVersionDeployments({
+          released: undefined,
+        });
+        assert.equal(result, BlastRouterProxyFactory_v2_1_0_0);
+        // NB: Add older old versions here
+        [RouterProxyFactory_v2_0_1_0, RouterProxyFactory_v2_0_0_b1_0].forEach(
+          (version) => {
+            assert.notEqual(result, version);
+          },
+        );
+      });
+    });
+
+    describe("getAllRouterProxyFactoryVersionDeploymentsPerNetwork", () => {
+      it("should return all deployments grouped by network", () => {
+        expect(
+          getAllRouterProxyFactoryVersionDeploymentsPerNetwork({
+            released: undefined,
+          }),
+        ).to.deep.equal({
+          "80001": [
+            RouterProxyFactory_v2_0_1_0,
+            RouterProxyFactory_v2_0_0_b1_0,
+          ],
+          "168587773": [BlastRouterProxyFactory_v2_1_0_0],
+        });
+      });
+    });
+
+    describe("getLatestRouterProxyFactoryPerNetwork", () => {
+      it("should return the latest deployment for each network", () => {
+        expect(
+          getLatestRouterProxyFactoryPerNetwork({ released: undefined }),
+        ).to.deep.equal({
+          "80001": firstVersionDeploymentsToVersionNetworkDeployment(
+            RouterProxyFactory_v2_0_1_0,
+            "80001",
+          ),
+          "168587773": firstVersionDeploymentsToVersionNetworkDeployment(
+            BlastRouterProxyFactory_v2_1_0_0,
+            "168587773",
+          ),
+        });
+      });
+    });
+  });
+
+  describe("SimpleAaveLogic contract", () => {
+    describe("getSimpleAaveLogicVersionDeployments", () => {
+      it("should find the latest deployment first", () => {
+        const result = getSimpleAaveLogicVersionDeployments({
+          released: undefined,
+        });
+        assert.equal(result, SimpleAaveLogic_v2_0_1_0);
+        // NB: Add older old versions here
+        [SimpleAaveLogic_v2_0_0_b1_0].forEach((version) => {
+          assert.notEqual(result, version);
+        });
+      });
+    });
+
+    describe("getAllSimpleAaveLogicVersionDeploymentsPerNetwork", () => {
+      it("should return all deployments grouped by network", () => {
+        expect(
+          getAllSimpleAaveLogicVersionDeploymentsPerNetwork({
+            released: undefined,
+          }),
+        ).to.deep.equal({
+          "80001": [SimpleAaveLogic_v2_0_1_0, SimpleAaveLogic_v2_0_0_b1_0],
+        });
+      });
+    });
+
+    describe("getLatestSimpleAaveLogicPerNetwork", () => {
+      it("should return the latest deployment for each network", () => {
+        expect(
+          getLatestSimpleAaveLogicPerNetwork({ released: undefined }),
+        ).to.deep.equal({
+          "80001": firstVersionDeploymentsToVersionNetworkDeployment(
+            SimpleAaveLogic_v2_0_1_0,
+            "80001",
+          ),
+        });
+      });
+    });
+  });
+
+  describe("MangroveAmplifier contract", () => {
+    describe("getMangroveAmplifierVersionDeployments", () => {
+      it("should find the latest deployment first", () => {
+        const result = getMangroveAmplifierVersionDeployments({
+          released: undefined,
+        });
+        assert.equal(result, BlastMangroveAmplifier_v2_1_0_0);
+        // NB: Add older old versions here
+        [MangroveAmplifier_v2_0_1_0, MangroveAmplifier_v2_0_0_b1_0].forEach(
+          (version) => {
+            assert.notEqual(result, version);
+          },
+        );
+      });
+    });
+
+    describe("getAllMangroveAmplifierVersionDeploymentsPerNetwork", () => {
+      it("should return all deployments grouped by network", () => {
+        expect(
+          getAllMangroveAmplifierVersionDeploymentsPerNetwork({
+            released: undefined,
+          }),
+        ).to.deep.equal({
+          "80001": [MangroveAmplifier_v2_0_1_0, MangroveAmplifier_v2_0_0_b1_0],
+          "168587773": [BlastMangroveAmplifier_v2_1_0_0],
+        });
+      });
+    });
+
+    describe("getLatestMangroveAmplifierPerNetwork", () => {
+      it("should return the latest deployment for each network", () => {
+        expect(
+          getLatestMangroveAmplifierPerNetwork({ released: undefined }),
+        ).to.deep.equal({
+          "80001": firstVersionDeploymentsToVersionNetworkDeployment(
+            MangroveAmplifier_v2_0_1_0,
+            "80001",
+          ),
+          "168587773": firstVersionDeploymentsToVersionNetworkDeployment(
+            BlastMangroveAmplifier_v2_1_0_0,
+            "168587773",
           ),
         });
       });
@@ -281,28 +510,40 @@ describe("strats.ts", () => {
             "80001",
           ),
           aaveKandelSeeder: firstVersionDeploymentsToVersionNetworkDeployment(
-            AaveKandelSeeder_v1_0_0,
+            AaveKandelSeeder_v2_0_1_0,
             "80001",
           ),
           aavePooledRouter: firstVersionDeploymentsToVersionNetworkDeployment(
-            AavePooledRouter_v1_0_0,
+            AavePooledRouter_v2_0_1_0,
             "80001",
           ),
           kandelLib: firstVersionDeploymentsToVersionNetworkDeployment(
-            KandelLib_v1_0_0,
+            KandelLib_v2_0_1_0,
             "80001",
           ),
           kandelSeeder: firstVersionDeploymentsToVersionNetworkDeployment(
-            KandelSeeder_v1_0_0,
+            KandelSeeder_v2_0_1_0,
             "80001",
           ),
           mangroveOrderRouter:
             firstVersionDeploymentsToVersionNetworkDeployment(
-              MangroveOrderRouter_v1_0_0,
+              MangroveOrderRouter_v2_0_1_0,
               "80001",
             ),
           mangroveOrder: firstVersionDeploymentsToVersionNetworkDeployment(
-            MangroveOrder_v1_0_0,
+            MangroveOrder_v2_0_1_0,
+            "80001",
+          ),
+          routerProxyFactory: firstVersionDeploymentsToVersionNetworkDeployment(
+            RouterProxyFactory_v2_0_1_0,
+            "80001",
+          ),
+          simpleAaveLogic: firstVersionDeploymentsToVersionNetworkDeployment(
+            SimpleAaveLogic_v2_0_1_0,
+            "80001",
+          ),
+          mangroveAmplifier: firstVersionDeploymentsToVersionNetworkDeployment(
+            MangroveAmplifier_v2_0_1_0,
             "80001",
           ),
         },
@@ -317,18 +558,43 @@ describe("strats.ts", () => {
           kandelSeeder: undefined,
           mangroveOrderRouter: undefined,
           mangroveOrder: undefined,
+          routerProxyFactory: undefined,
+          simpleAaveLogic: undefined,
+          mangroveAmplifier: undefined,
         },
         "168587773": {
           mangrove: firstVersionDeploymentsToVersionNetworkDeployment(
-            Mangrove_v2_0_1,
+            BlastMangrove_v2_1_0_0,
             "168587773",
           ),
           aaveKandelSeeder: undefined,
           aavePooledRouter: undefined,
-          kandelLib: undefined,
-          kandelSeeder: undefined,
-          mangroveOrderRouter: undefined,
-          mangroveOrder: undefined,
+          kandelLib: firstVersionDeploymentsToVersionNetworkDeployment(
+            KandelLib_v2_0_1_0,
+            "168587773",
+          ),
+          kandelSeeder: firstVersionDeploymentsToVersionNetworkDeployment(
+            KandelSeeder_v2_0_1_0,
+            "168587773",
+          ),
+          mangroveOrderRouter:
+            firstVersionDeploymentsToVersionNetworkDeployment(
+              BlastMangroveOrderRouter_v2_1_0_0,
+              "168587773",
+            ),
+          mangroveOrder: firstVersionDeploymentsToVersionNetworkDeployment(
+            BlastMangroveOrder_v2_1_0_0,
+            "168587773",
+          ),
+          routerProxyFactory: firstVersionDeploymentsToVersionNetworkDeployment(
+            BlastRouterProxyFactory_v2_1_0_0,
+            "168587773",
+          ),
+          simpleAaveLogic: undefined,
+          mangroveAmplifier: firstVersionDeploymentsToVersionNetworkDeployment(
+            BlastMangroveAmplifier_v2_1_0_0,
+            "168587773",
+          ),
         },
       });
     });